requirements.txt
phidata
langchain
faiss-cpu
numpy
pandas
scikit-learn
torch
transformers
tqdm
flask
uvicorn
fastapi
python-dotenv
yfinance
packaging
groq
duckduckgo-search
openai
streamlit
